Reworked class to improve encapsulation
Code implements the logic to configure what files will be merged into an aggregated schematron
At present it is works with property files. MergeConfigure work should not be specific in the use use of Property files. These changes set the groundwork for allowing the specification of Property Files, Extended Property FIles, or even JSON files to specify the templates to be merged.